+++ packages/react-native-messaging-core/src/NablaMessagingClient/NativeModules/NablaMessagingClientModule.ts
@@ -4,31 +4,38 @@ import { ConversationId, MessageId } from '../../types';
 
 const LINKING_ERROR =
   `The package '@nabla/react-native-messaging-core' doesn't seem to be linked. Make sure: \n\n` +
-  Platform.select({ ios: "- You have run 'pod install'\n", default: '' }) +
+  Platform.select({ ios: '- You have run \'pod install\'\n', default: '' }) +
   '- You rebuilt the app after installing the package\n' +
   '- You are not using Expo managed workflow\n';
+
+type Callback<T> = (
+  error: NativeError | undefined,
+  result: T | undefined,
+) => void;
 
 interface NablaMessagingClientModule extends NativeModule {
   createConversation(
     title: String | undefined,
     providerIds: String[] | undefined,
-    callback: (
-      error: NativeError | undefined,
-      conversationId: ConversationId | undefined,
-    ) => void,
+    callback: Callback<ConversationId>
   ): void;
 
   sendMessage(
     input: any,
     conversationId: ConversationId,
     replyTo: MessageId | undefined,
-    callback: (error: NativeError | undefined) => void,
+    callback: Callback<void>,
   ): void;
 
   deleteMessage(
     messageId: MessageId,
     conversationId: ConversationId,
-    callback: (error: NativeError | undefined) => void,
+    callback: Callback<void>,
+  ): void;
+
+  markConversationAsSeen(
+    conversationId: ConversationId,
+    callback: Callback<void>,
   ): void;
 }
 
@@ -36,10 +43,10 @@ export const nablaMessagingClientModule: NablaMessagingClientModule =
   NativeModules.NablaMessagingClientModule
     ? NativeModules.NablaMessagingClientModule
     : new Proxy(
-        {},
-        {
-          get() {
-            throw new Error(LINKING_ERROR);
-          },
+      {},
+      {
+        get() {
+          throw new Error(LINKING_ERROR);
         },
-      );
+      },
+    );